+++ WordPress/Tests/Security/ValidatedSanitizedInputUnitTest.php
@@ -7,7 +7,7 @@
  * @license https://opensource.org/licenses/MIT MIT
  */
 
-namespace WordPress\Tests\VIP;
+namespace WordPress\Tests\Security;
 
 use PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest;
 
@@ -18,9 +18,7 @@ use PHP_CodeSniffer\Tests\Standards\AbstractSniffUnitTest;
  *
  * @since   0.3.0
  * @since   0.13.0 Class name changed: this class is now namespaced.
- * @since   0.15.0 The sniff has been deprecated. This unit test file now
- *                 only tests that the deprecation warnings are correctly thrown
- *                 and that the sniff falls through to the parent correctly.
+ * @since   0.15.0 This sniff has been moved from the `VIP` category to the `Security` category.
  */
 class ValidatedSanitizedInputUnitTest extends AbstractSniffUnitTest {
 
@@ -31,7 +29,31 @@ class ValidatedSanitizedInputUnitTest extends AbstractSniffUnitTest {
 	 */
 	public function getErrorList() {
 		return array(
-			4 => 3,
+			5   => 3,
+			7   => 1,
+			10  => 1,
+			20  => 1,
+			33  => 3,
+			65  => 1,
+			79  => 1,
+			80  => 1,
+			81  => 1,
+			82  => 1,
+			85  => 1,
+			90  => 1,
+			93  => 1,
+			96  => 1,
+			100 => 2,
+			101 => 1,
+			104 => 2,
+			105 => 1,
+			114 => 2,
+			121 => 1,
+			132 => 1,
+			137 => 1,
+			138 => 1,
+			150 => 2,
+			160 => 2,
 		);
 
 	}
@@ -42,9 +64,7 @@ class ValidatedSanitizedInputUnitTest extends AbstractSniffUnitTest {
 	 * @return array <int line number> => <int number of warnings>
 	 */
 	public function getWarningList() {
-		return array(
-			1 => 2,
-		);
+		return array();
 
 	}
